requirements.txt
asgiref==3.7.2
click==8.1.7
Django==4.1.13
djangorestframework==3.15.0
djongo==1.3.6
dnspython==2.6.1
gunicorn==21.2.0
joblib==1.3.2
nltk==3.8.1
packaging==24.0
pymongo==3.12.1
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pytz==2024.1
regex==2023.12.25
sqlparse==0.2.4
tqdm==4.66.2
typing_extensions==4.10.0
